Move RocksDBMethods implementations into separate files.
