+++ arangod/RocksDBEngine/RocksDBBuilderIndex.cpp
@@ -27,6 +27,8 @@
 #include "Basics/VelocyPackHelper.h"
 #include "Basics/application-exit.h"
 #include "Containers/HashSet.h"
+#include "RocksDBEngine/Methods/RocksDBBatchedMethods.h"
+#include "RocksDBEngine/Methods/RocksDBBatchedWithIndexMethods.h"
 #include "RocksDBEngine/RocksDBCollection.h"
 #include "RocksDBEngine/RocksDBColumnFamilyManager.h"
 #include "RocksDBEngine/RocksDBCommon.h"
@@ -37,7 +39,6 @@
 #include "RocksDBEngine/RocksDBTransactionState.h"
 #include "StorageEngine/EngineSelectorFeature.h"
 #include "Transaction/StandaloneContext.h"
-#include "Utils/SingleCollectionTransaction.h"
 #include "VocBase/LogicalCollection.h"
 #include "VocBase/ticks.h"
 
@@ -151,9 +152,9 @@ Result RocksDBBuilderIndex::remove(transaction::Methods& trx, RocksDBMethods* mt
 }
 
 // fast mode assuming exclusive access locked from outside
-template <typename WriteBatchType, typename MethodsType, bool foreground>
-static arangodb::Result fillIndex(rocksdb::DB* rootDB, RocksDBIndex& ridx,
-                                  WriteBatchType& batch, rocksdb::Snapshot const* snap) {
+template <bool foreground>
+static arangodb::Result fillIndex(rocksdb::DB* rootDB, RocksDBIndex& ridx, RocksDBMethods& batched, rocksdb::WriteBatchBase& batch,
+                                  rocksdb::Snapshot const* snap) {
   // fillindex can be non transactional, we just need to clean up
   TRI_ASSERT(rootDB != nullptr);
 
@@ -192,8 +193,6 @@ static arangodb::Result fillIndex(rocksdb::DB* rootDB, RocksDBIndex& ridx,
   uint64_t numDocsWritten = 0;
   auto state = RocksDBTransactionState::toState(&trx);
   RocksDBTransactionCollection* trxColl = trx.resolveTrxCollection();
-  // write batch will be reset every x documents
-  MethodsType batched(state, &batch);
   
   auto commitLambda = [&] {
     if (batch.GetWriteBatch()->Count() > 0) {
@@ -292,14 +291,14 @@ arangodb::Result RocksDBBuilderIndex::fillIndexForeground() {
     // unique index. we need to keep track of all our changes because we need to
     // avoid duplicate index keys. must therefore use a WriteBatchWithIndex
     rocksdb::WriteBatchWithIndex batch(cmp, 32 * 1024 * 1024);
-    res = ::fillIndex<rocksdb::WriteBatchWithIndex, RocksDBBatchedWithIndexMethods, true>(
-        db, *internal, batch, snap);
+    RocksDBBatchedWithIndexMethods methods(engine.db(), &batch);
+    res = ::fillIndex<true>(db, *internal, methods, batch, snap);
   } else {
     // non-unique index. all index keys will be unique anyway because they
     // contain the document id we can therefore get away with a cheap WriteBatch
     rocksdb::WriteBatch batch(32 * 1024 * 1024);
-    res = ::fillIndex<rocksdb::WriteBatch, RocksDBBatchedMethods, true>(db, *internal,
-                                                                        batch, snap);
+    RocksDBBatchedMethods methods(&batch);
+    res = ::fillIndex<true>(db, *internal, methods, batch, snap);
   }
 
   return res;
@@ -307,9 +306,8 @@ arangodb::Result RocksDBBuilderIndex::fillIndexForeground() {
 
 namespace {
 
-template <typename MethodsType>
 struct ReplayHandler final : public rocksdb::WriteBatch::Handler {
-  ReplayHandler(uint64_t oid, RocksDBIndex& idx, transaction::Methods& trx, MethodsType* methods)
+  ReplayHandler(uint64_t oid, RocksDBIndex& idx, transaction::Methods& trx, RocksDBMethods* methods)
       : _objectId(oid), _index(idx), _trx(trx), _methods(methods) {}
 
   bool Continue() override {
@@ -432,7 +430,7 @@ struct ReplayHandler final : public rocksdb::WriteBatch::Handler {
   const uint64_t _objectId;  /// collection objectID
   RocksDBIndex& _index;      /// the index to use
   transaction::Methods& _trx;
-  MethodsType* _methods;  /// methods to fill
+  RocksDBMethods* _methods;  /// methods to fill
   OperationOptions const _options;
 
   rocksdb::SequenceNumber _startSequence;
@@ -442,8 +440,7 @@ struct ReplayHandler final : public rocksdb::WriteBatch::Handler {
   uint64_t _lastObjectID = 0;
 };
 
-template <typename WriteBatchType, typename MethodsType>
-Result catchup(rocksdb::DB* rootDB, RocksDBIndex& ridx, WriteBatchType& wb,
+Result catchup(rocksdb::DB* rootDB, RocksDBIndex& ridx, RocksDBMethods& batched, rocksdb::WriteBatchBase& wb,
                AccessMode::Type mode, rocksdb::SequenceNumber startingFrom,
                rocksdb::SequenceNumber& lastScannedTick, uint64_t& numScanned,
                bool haveExclusiveAccess) {
@@ -463,10 +460,7 @@ Result catchup(rocksdb::DB* rootDB, RocksDBIndex& ridx, WriteBatchType& wb,
 
   TRI_ASSERT(rootDB != nullptr);
 
-  // write batch will be reset every x documents
-  MethodsType batched(state, &wb);
-
-  ReplayHandler<MethodsType> replay(rcoll->objectId(), ridx, trx, &batched);
+  ReplayHandler replay(rcoll->objectId(), ridx, trx, &batched);
 
   std::unique_ptr<rocksdb::TransactionLogIterator> iterator;  // reader();
   // no need verifying the WAL contents
@@ -592,14 +586,14 @@ arangodb::Result RocksDBBuilderIndex::fillIndexBackground(Locker& locker) {
     // unique index. we need to keep track of all our changes because we need to
     // avoid duplicate index keys. must therefore use a WriteBatchWithIndex
     rocksdb::WriteBatchWithIndex batch(cmp, 32 * 1024 * 1024);
-    res = ::fillIndex<rocksdb::WriteBatchWithIndex, RocksDBBatchedWithIndexMethods, false>(
-        db, *internal, batch, snap);
+    RocksDBBatchedWithIndexMethods methods(engine.db(), &batch);
+    res = ::fillIndex<false>(db, *internal, methods, batch, snap);
   } else {
     // non-unique index. all index keys will be unique anyway because they
     // contain the document id we can therefore get away with a cheap WriteBatch
     rocksdb::WriteBatch batch(32 * 1024 * 1024);
-    res = ::fillIndex<rocksdb::WriteBatch, RocksDBBatchedMethods, false>(db, *internal,
-                                                                         batch, snap);
+    RocksDBBatchedMethods methods(&batch);
+    res = ::fillIndex<false>(db, *internal, methods, batch, snap);
   }
   
   if (res.fail()) {
@@ -620,17 +614,16 @@ arangodb::Result RocksDBBuilderIndex::fillIndexBackground(Locker& locker) {
       // unique index. we need to keep track of all our changes because we need
       // to avoid duplicate index keys. must therefore use a WriteBatchWithIndex
       rocksdb::WriteBatchWithIndex batch(cmp, 32 * 1024 * 1024);
-      res = ::catchup<rocksdb::WriteBatchWithIndex, RocksDBBatchedWithIndexMethods>(
-          db, *internal, batch, AccessMode::Type::WRITE, scanFrom, lastScanned,
-          numScanned, false);
+      RocksDBBatchedWithIndexMethods methods(engine.db(), &batch);
+      res = ::catchup(db, *internal, methods, batch, AccessMode::Type::WRITE, scanFrom, lastScanned,
+        numScanned, false);
     } else {
       // non-unique index. all index keys will be unique anyway because they
       // contain the document id we can therefore get away with a cheap WriteBatch
       rocksdb::WriteBatch batch(32 * 1024 * 1024);
-      res = ::catchup<rocksdb::WriteBatch, RocksDBBatchedMethods>(db, *internal, batch,
-                                                                  AccessMode::Type::WRITE,
-                                                                  scanFrom, lastScanned,
-                                                                  numScanned, false);
+      RocksDBBatchedMethods methods(&batch);
+      res = ::catchup(db, *internal, methods, batch, AccessMode::Type::WRITE, scanFrom, lastScanned,
+        numScanned, false);
     }
 
     if (res.fail() && !res.is(TRI_ERROR_ARANGO_TRY_AGAIN)) {
@@ -652,17 +645,16 @@ arangodb::Result RocksDBBuilderIndex::fillIndexBackground(Locker& locker) {
     // unique index. we need to keep track of all our changes because we need to
     // avoid duplicate index keys. must therefore use a WriteBatchWithIndex
     rocksdb::WriteBatchWithIndex batch(cmp, 32 * 1024 * 1024);
-    res = ::catchup<rocksdb::WriteBatchWithIndex, RocksDBBatchedWithIndexMethods>(
-        db, *internal, batch, AccessMode::Type::EXCLUSIVE, scanFrom,
-        lastScanned, numScanned, true);
+    RocksDBBatchedWithIndexMethods methods(engine.db(), &batch);
+    res = ::catchup(db, *internal, methods, batch, AccessMode::Type::EXCLUSIVE, scanFrom,
+      lastScanned, numScanned, true);
   } else {
     // non-unique index. all index keys will be unique anyway because they
     // contain the document id we can therefore get away with a cheap WriteBatch
     rocksdb::WriteBatch batch(32 * 1024 * 1024);
-    res = ::catchup<rocksdb::WriteBatch, RocksDBBatchedMethods>(db, *internal, batch,
-                                                                AccessMode::Type::EXCLUSIVE,
-                                                                scanFrom, lastScanned,
-                                                                numScanned, true);
+    RocksDBBatchedMethods methods(&batch);
+    res = ::catchup(db, *internal, methods, batch,AccessMode::Type::EXCLUSIVE, scanFrom,
+      lastScanned, numScanned, true);
   }
 
   return res;
