+++ arangod/RocksDBEngine/RocksDBMethods.h
@@ -26,13 +26,7 @@
 #include "RocksDBEngine/RocksDBCommon.h"
 
 namespace rocksdb {
-class Transaction;
-class Slice;
-class Iterator;
-class TransactionDB;
-class WriteBatch;
-class WriteBatchWithIndex;
-struct ReadOptions;
+ class Slice;
 }  // namespace rocksdb
 
 namespace arangodb {
@@ -91,149 +85,6 @@ class RocksDBMethods {
 
  protected:
   RocksDBTransactionState* _state;
-};
-
-// only implements GET and NewIterator
-class RocksDBReadOnlyMethods final : public RocksDBMethods {
- public:
-  explicit RocksDBReadOnlyMethods(RocksDBTransactionState* state);
-
-  rocksdb::Status Get(rocksdb::ColumnFamilyHandle*, rocksdb::Slice const& key,
-                      rocksdb::PinnableSlice* val) override;
-  rocksdb::Status GetForUpdate(rocksdb::ColumnFamilyHandle*,
-                               rocksdb::Slice const&,
-                               rocksdb::PinnableSlice*) override;
-  rocksdb::Status Put(rocksdb::ColumnFamilyHandle*, RocksDBKey const& key,
-                      rocksdb::Slice const& val, bool assume_tracked) override;
-  rocksdb::Status PutUntracked(rocksdb::ColumnFamilyHandle*, RocksDBKey const& key,
-                               rocksdb::Slice const& val) override;
-  rocksdb::Status Delete(rocksdb::ColumnFamilyHandle*, RocksDBKey const& key) override;
-  rocksdb::Status SingleDelete(rocksdb::ColumnFamilyHandle*, RocksDBKey const&) override;
-  void PutLogData(rocksdb::Slice const&) override;
-
-  std::unique_ptr<rocksdb::Iterator> NewIterator(rocksdb::ReadOptions const&,
-                                                 rocksdb::ColumnFamilyHandle*) override;
-
-  void SetSavePoint() override {}
-  rocksdb::Status RollbackToSavePoint() override {
-    return rocksdb::Status::OK();
-  }
-  rocksdb::Status RollbackToWriteBatchSavePoint() override {
-    // simply relay to the general method (which in this derived class does nothing)
-    return RollbackToSavePoint();
-  }
-  void PopSavePoint() override {}
-
- private:
-  rocksdb::TransactionDB* _db;
-};
-
-/// transaction wrapper, uses the current rocksdb transaction
-class RocksDBTrxMethods : public RocksDBMethods {
- public:
-  explicit RocksDBTrxMethods(RocksDBTransactionState* state);
-
-  virtual bool isIndexingDisabled() const override{ return _indexingDisabled; }
-
-  /// @brief returns true if indexing was disabled by this call
-  bool DisableIndexing() override;
-
-  bool EnableIndexing() override;
-
-  rocksdb::Status Get(rocksdb::ColumnFamilyHandle*, rocksdb::Slice const& key,
-                      rocksdb::PinnableSlice* val) override;
-  rocksdb::Status GetForUpdate(rocksdb::ColumnFamilyHandle*,
-                               rocksdb::Slice const&,
-                               rocksdb::PinnableSlice*) override;
-  rocksdb::Status Put(rocksdb::ColumnFamilyHandle*, RocksDBKey const& key,
-                      rocksdb::Slice const& val, bool assume_tracked) override;
-  rocksdb::Status PutUntracked(rocksdb::ColumnFamilyHandle*, RocksDBKey const& key,
-                               rocksdb::Slice const& val) override;
-  rocksdb::Status Delete(rocksdb::ColumnFamilyHandle*, RocksDBKey const& key) override;
-  rocksdb::Status SingleDelete(rocksdb::ColumnFamilyHandle*, RocksDBKey const&) override;
-  void PutLogData(rocksdb::Slice const&) override;
-
-  std::unique_ptr<rocksdb::Iterator> NewIterator(rocksdb::ReadOptions const&,
-                                                 rocksdb::ColumnFamilyHandle*) override;
-
-  void SetSavePoint() override;
-  rocksdb::Status RollbackToSavePoint() override;
-  rocksdb::Status RollbackToWriteBatchSavePoint() override;
-  void PopSavePoint() override;
-
-  bool _indexingDisabled;
-};
-
-
-/// wraps a writebatch - non transactional
-class RocksDBBatchedMethods final : public RocksDBMethods {
- public:
-  RocksDBBatchedMethods(RocksDBTransactionState*, rocksdb::WriteBatch*);
-
-  rocksdb::Status Get(rocksdb::ColumnFamilyHandle*, rocksdb::Slice const& key,
-                      rocksdb::PinnableSlice* val) override;
-  rocksdb::Status GetForUpdate(rocksdb::ColumnFamilyHandle*,
-                               rocksdb::Slice const&,
-                               rocksdb::PinnableSlice*) override;
-  rocksdb::Status Put(rocksdb::ColumnFamilyHandle*, RocksDBKey const& key,
-                      rocksdb::Slice const& val, bool assume_tracked) override;
-  rocksdb::Status PutUntracked(rocksdb::ColumnFamilyHandle*, RocksDBKey const& key,
-                               rocksdb::Slice const& val) override;
-  rocksdb::Status Delete(rocksdb::ColumnFamilyHandle*, RocksDBKey const& key) override;
-  rocksdb::Status SingleDelete(rocksdb::ColumnFamilyHandle*, RocksDBKey const&) override;
-  void PutLogData(rocksdb::Slice const&) override;
-
-  std::unique_ptr<rocksdb::Iterator> NewIterator(rocksdb::ReadOptions const&,
-                                                 rocksdb::ColumnFamilyHandle*) override;
-
-  void SetSavePoint() override {}
-  rocksdb::Status RollbackToSavePoint() override {
-    return rocksdb::Status::OK();
-  }
-  rocksdb::Status RollbackToWriteBatchSavePoint() override {
-    // simply relay to the general method (which in this derived class does nothing)
-    return RollbackToSavePoint();
-  }
-  void PopSavePoint() override {}
-
- private:
-  rocksdb::WriteBatch* _wb;
-};
-
-/// wraps a writebatch with index - non transactional
-class RocksDBBatchedWithIndexMethods final : public RocksDBMethods {
- public:
-  RocksDBBatchedWithIndexMethods(RocksDBTransactionState*, rocksdb::WriteBatchWithIndex*);
-
-  rocksdb::Status Get(rocksdb::ColumnFamilyHandle*, rocksdb::Slice const& key,
-                      rocksdb::PinnableSlice* val) override;
-  rocksdb::Status GetForUpdate(rocksdb::ColumnFamilyHandle*,
-                               rocksdb::Slice const&,
-                               rocksdb::PinnableSlice*) override;
-  rocksdb::Status Put(rocksdb::ColumnFamilyHandle*, RocksDBKey const& key,
-                      rocksdb::Slice const& val, bool assume_tracked) override;
-  rocksdb::Status PutUntracked(rocksdb::ColumnFamilyHandle*, RocksDBKey const& key,
-                               rocksdb::Slice const& val) override;
-  rocksdb::Status Delete(rocksdb::ColumnFamilyHandle*, RocksDBKey const& key) override;
-  rocksdb::Status SingleDelete(rocksdb::ColumnFamilyHandle*, RocksDBKey const&) override;
-  void PutLogData(rocksdb::Slice const&) override;
-
-  std::unique_ptr<rocksdb::Iterator> NewIterator(rocksdb::ReadOptions const&,
-                                                 rocksdb::ColumnFamilyHandle*) override;
-
-  void SetSavePoint() override {}
-  rocksdb::Status RollbackToSavePoint() override {
-    return rocksdb::Status::OK();
-  }
-  rocksdb::Status RollbackToWriteBatchSavePoint() override {
-    // simply relay to the general method (which in this derived class does nothing)
-    return RollbackToSavePoint();
-  }
-  void PopSavePoint() override {}
-
- private:
-  rocksdb::TransactionDB* _db;
-  rocksdb::WriteBatchWithIndex* _wb;
 };
 
 // INDEXING MAY ONLY BE DISABLED IN TOPLEVEL AQL TRANSACTIONS
